Submission to Assignment
Added Powerups and implemented the audio I added a while back.

KNOWN BUGS:
For some reason, the +5 point powerup has lost the ability to destroy itself after implementing audio, which makes zero sense as it should still be able to explode. The Nuke Powerup does this on a 50/50 basis. I am unsure why this occurs, though it may be a strange thing with Unity Timing.